adicionada constante de dominio e outras melhorias
