+++ cadastroequipamento.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/css/cadastro.css">
+    <link rel="stylesheet" href="/css/cadastroequipamento.css">
     <title> Castastro Equipamento</title>
 </head>
 <body>
@@ -38,11 +38,6 @@
                                 <option value="ALGEMA">ALGEMA</option>
                                 <option value="ARMA_LETAL">ARMA LETAL</option>
                                 <option value="ARMA_MENOS_LETAL">ARMA MENOS LETAL</option>
-                                <!-- <optgroup label="ARMAS DE FOGO">
-                                    <option value="ARMA_9">9MM</option>
-                                    <option value="ARMA_380">380</option>
-                                    <option value="ARMA_12">CALIBRE 12</option>
-                                </optgroup> -->
                                 <option value="ESPARGIDOR">ESPARGIDOR</option>
                                 <option value="MUNICAO">MUNIÇÃO</option>
                                 <option value="GRANADA">GRANADA</option>
