+++ css/cadastroequipamento.css
@@ -1,0 +1,192 @@
+*, *:before, *:after {
+    box-sizing: border-box;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+}
+  
+body {
+    background: linear-gradient(to right, #232427 10%, #194471 150%);
+    font-size: 12px;
+}
+  
+body, button, input {
+font-family: 'Montserrat', sans-serif;
+font-weight: 700;
+letter-spacing: 1.4px;
+}
+  
+.background {
+display: flex;
+min-height: 100vh;
+}
+  
+.container {
+flex: 0 1 700px;
+margin: auto;
+padding: 10px;
+}
+  
+.screen {
+position: relative;
+background: #3e3e3e;
+border-radius: 15px;
+}
+  
+.screen:after {
+content: '';
+display: block;
+position: absolute;
+top: 0;
+left: 20px;
+right: 20px;
+bottom: 0;
+border-radius: 15px;
+box-shadow: 0 20px 40px rgba(0, 0, 0, .4);
+z-index: -1;
+}
+  
+.screen-header {
+display: flex;
+align-items: center;
+padding: 10px 20px;
+background: #4d4d4f;
+border-top-left-radius: 15px;
+border-top-right-radius: 15px;
+}
+  
+.screen-header-left {
+margin-right: auto;
+}
+  
+.screen-header-right {
+display: flex;
+}
+  
+.screen-header-button {
+display: inline-block;
+width: 8px;
+height: 8px;
+margin-right: 3px;
+border-radius: 8px;
+background: white;
+}
+  
+.screen-header-ellipsis {
+width: 3px;
+height: 3px;
+margin-left: 2px;
+border-radius: 8px;
+background: #999;
+}
+  
+.screen-body {
+display: flex;
+flex-direction: column;
+}
+  
+.screen-body-item {
+padding: 50px 50px;
+}
+  
+.screen-body-item.left {
+display: flex;
+flex-direction: column;
+padding: 50px 50px 0px;
+}
+  
+.app-title {
+display: flex;
+flex-direction: column;
+position: relative;
+color: #B9E0E6;
+font-size: 30px;
+}
+  
+.app-form-group {
+margin-bottom: 15px;
+}
+  
+.app-form-group.message {
+margin-top: 40px;
+}
+  
+.app-form-group.buttons {
+margin-bottom: 0;
+text-align: right;
+}
+  
+.app-form-control {
+width: 100%;
+padding: 10px 0;
+background: none;
+border: none;
+border-bottom: 1px solid #666;
+color: #ddd;
+font-size: 14px;
+outline: none;
+transition: border-color .2s;
+}
+  
+.app-form-control::placeholder {
+color: #666;
+}
+
+.app-form-control:focus {
+border-bottom-color: #ddd;
+}
+  
+.app-form-button {
+border: none;
+color: #2196f3;
+font-size: 14px;
+cursor: pointer;
+outline: none;
+left: 90px;
+display: inline-block;
+padding: 10px 15px;
+background: #2196f3;
+border-radius: 6px;
+text-decoration: none;
+color: white;
+margin: 20px 5px 0px;
+box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
+transition: .8s;
+}
+  
+.app-form-button:hover {
+box-shadow: 0 10px 20px rgba(0, 0, 0, 0.6);
+background: #B0E0E6;
+color: #000;
+}
+  
+@media screen and (max-width: 520px) {
+    .screen-body {
+        flex-direction: column;
+    }
+    
+    .screen-body-item.left {
+        margin-bottom: 30px;
+    }
+    
+    .app-title {
+        flex-direction: row;
+    }
+    
+    .app-title span {
+        margin-right: 12px;
+    }
+    
+    .app-title:after {
+        display: none;
+    }
+}
+  
+@media screen and (max-width: 600px) {
+    .screen-body {
+        padding: 40px;
+    }
+
+    .screen-body-item {
+        padding: 0;
+    }
+}
