+++ index.html
@@ -191,6 +191,6 @@
         </main>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script type="module" src="/js/index.js"></script>
+    <script src="/js/index.js"></script>
 </body>
 </html>
