+++ js/index.js
@@ -1,9 +1,30 @@
+//const dominio = 'http://localhost:8080';//README
+const dominio = 'https://gcmsystem.up.railway.app';
 let config = {
     headers: {'Authorization': 'Bearer ' + sessionStorage.getItem('token')}
 }
-//let dominio = 'https://gcmsystem.up.railway.app'; //RAILWAY
-//let dominio = 'http://localhost:8080';
-import { dominio } from "../modules/dominio.js";
+//API busca total de GCMs Cadastrados no BD por padrão método GET
+axios.get(dominio+'/gcm/total', config)
+    .then(response => {totalGcm.textContent = JSON.stringify(response.data)})
+    .catch(function(error){
+        console.log("Erro "+error.response.status+": "+error.response.data)
+        window.location.href = "/login.html";
+    }
+)
+//API busca total de Equipamentos Cadastrados no BD por padrão método GET
+axios.get(dominio+'/equipment/total', config)
+    .then(response => {totalEquip.textContent = JSON.stringify(response.data)})
+    .catch(function(error){
+        console.log("Busca total equipamentos falhou. Motivo:")
+        console.log("Erro "+error.response.status+": "+error.response.data)
+        window.location.href = "/login.html";
+    }
+)
+//exclui token de acesso após clicar em sair
+function logout(){
+    sessionStorage.removeItem('token');
+}
+
 //******************* API busca CEP ***************************************//
 // function buscarCep(){
 //     let recebeCep =  document.getElementById("cepBuscar").value
@@ -43,26 +64,3 @@ import { dominio } from "../modules/dominio.js";
 //         })
 //         .catch(error=> console.log(error))
 // }
-//API busca total de GCMs Cadastrados no BD por padrão método GET
-axios.get(dominio+'/gcm/total', config)
-    .then(response => {totalGcm.textContent = JSON.stringify(response.data)})
-    .catch(function(error){
-        console.log("Erro "+error.response.status+": "+error.response.data)
-        window.location.href = "/login.html";
-    }
-)
-
-//API busca total de Equipamentos Cadastrados no BD por padrão método GET
-axios.get(dominio+'/equipment/total', config)
-    .then(response => {totalEquip.textContent = JSON.stringify(response.data)})
-    .catch(function(error){
-        console.log("Busca total equipamentos falhou. Motivo:")
-        console.log("Erro "+error.response.status+": "+error.response.data)
-        window.location.href = "/login.html";
-    }
-)
-
-//exclui token de acesso após clicar em sair
-function logout(){
-    sessionStorage.removeItem('token');
-}
