+++ js/preload.js
@@ -1,5 +1,6 @@
+// Apresenta semi-círculos girando enquanto a página carrega
 document.addEventListener("DOMContentLoaded", function() {
     setTimeout(function() {
         document.querySelector("body").classList.add("loaded");
-    }, 2500)
+    }, 2000)
 });
