+++ listagcms.html
@@ -22,6 +22,6 @@
   <main>
     <div id="container_mm"></div>
   </main>
-  <script type="module" src="/js/listagcms.js"></script>
+  <script src="/js/listagcms.js"></script>
 </body>
 </html>
