+++ login.html
@@ -30,6 +30,6 @@
             </div>
         </div>
     </div>
-    <script type="module" src="/js/login.js"></script>
+    <script src="/js/login.js"></script>
 </body>
 </html>
