+++ mostraequipamento.html
@@ -50,6 +50,7 @@
     <!-- Fim /modal solicita cartão de identificação -->
     <script src="/js/mostraequipamento.js"></script>
     <script src="/js/atribuirequipamento.js"></script>
+    <script src="/js/confirmpassword.js"></script>
     <!-- Script para gerar PDF -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/1.5.3/jspdf.debug.js" integrity="sha384-NaWTHo/8YCBYJ59830LTz/P4aQZK1sS0SneOgAvhsIl3zBu8r9RevNg5lHCHAuQ/" crossorigin="anonymous"></script>
     <!-- Script para enviar EMAIL -->
